data/datasets/cwru/requirements.txt
# Python dependencies for CWRU dataset pipeline
# Install: pip3 install -r requirements.txt

# MATLAB file reading
scipy>=1.10.0

# Numerical operations
numpy>=1.24.0

# CSV/DataFrame handling
pandas>=2.0.0

# Serial communication (for streaming)
pyserial>=3.5